Update patch set 1
Patch Set 1:

Well, this just enables and option that is kept disabled by default (config_proximityCheckOnWakeEnabledByDefault) and someone maybe could use it. I know I would.

I think Motorola improved the power button with the second gen, but it's not great on the first gen. Unlike for the volume rocker, my power button doesn't click (I need to press it in a specific way to make it sort of "click") and you just need a light touch to press it (depending on how you do it, don't know how to explain this). I'm sure I unintentionally pressed the power button while keeping my phone in my pocket and even managed to change song and unlock the screen.

The power button wasn't great when I got the phone and now it got worse.

Patch-set: 1