Update patch set 1
Patch Set 1:

> Is the phone unlocking even if the proximity sensor is covered when
 > you press-spam the power button (on falcon)? I remember this
 > happening on old CM12.1 days

Nope. Did that happen with the default timeout? I honestly don't remember. Here it happens if I lower the timeout [1]. The default timeout (300ms) is fine. Android will not wait for it to expire if it receives a value from the sensor early, so lower values won't make things faster (well, they could, but in an unwanted way).

[1] CT406 takes ~65ms to send the first value when activated while the sensor powered off and ~110ms when activated while the sensor is powered on, but not active.

Patch-set: 1