Update patch set 2
Patch Set 2: Code-Review-1

I never liked MotoDoze because for it to work the proximity sensor had to be on 24/7 (gestures). Or in the case of pocket mode, on when the screen is off.

Pocket mode is not needed any more because default ambient display has this function build in by default and it works now for our device (titan).

What I do want to implement is doze on pickup (kanged from hammerhead) I have it locally and will see if I get it to work during the upcoming days.

Patch-set: 2