Update patch set 2
Patch Set 2:

> I never liked MotoDoze because for it to work the proximity sensor
 > had to be on 24/7 (gestures). Or in the case of pocket mode, on
 > when the screen is off.
 > 

Proximity sensor is not on 24/7 for gestures neither, it is on only when the screen is off, just like with the pocket mode.

 > Pocket mode is not needed any more because default ambient display
 > has this function build in by default and it works now for our
 > device (titan).
 > 

That's not what we have on falcon, because of that i don't agree with the decision of removing it, pocket mode here clearily works differently from handwave and users may want to disable one of the two features.

Patch-set: 2