Update patch set 1
Patch Set 1:

> 3- Pocket mode is essentially 'doze_proximity_check_before_pulse' which is a fw/base overlay. It is on and works.

Lorenzo pointed out that it's actually a different feature. The one if fw/base prevents pulses if the proximity sensor is covered, pocket mode delays them.

 > Something else. From what I've seen pickup sensors are just emulated using an underlaying lowpower sensor like sig-motion or acel-sensor (hammerhead). I think something like this is more elegant. If I get some time I may try this.

Significant motion sensors and pickup sensors are most of the times virtual sensors obtained putting together the data coming from actual low-power sensors. The important thing is that these sensors must be able to work even when the CPU is off and wakes it when there's a meaningful event (e.g. significant motion).

LIS3DH (the accelerometer of falcon and peregrine) is capable of working independently on the CPU and wake it when certain conditions are met (http://review.cyanogenmod.org/#/c/124195/). Unfortunately the conditions for a significant motion event are different from those needed for a pickup event, so it's either one or the other, at least without the CPU.

Not sure about your Bosch sensor, which seems to be more featureful.
AFAIK hammerhead has a sensor hub, which allows to perform complex tasks without the need of the CPU.

Patch-set: 1