Update patch set 1
Patch Set 1:

I've just remembered about this. The lack of those files causes soft reboots when uninstalling apps. Not sure if there's a better way to do this or if it's better to go back to including the firmware in /system.

Not sure if this affect all the devices, I know that the radio images for falcon that were shipped with KitKat didn't include the keymaster files.

Patch-set: 1
Reviewer: Gabriele M <13897@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0