Update patch set 1
Abandoned

I don't really like this solution.

Only modems that came with Android 4.3 and 4.4.2 don't have the needed files, so I think this issue affects only falcon and there must be just a bunch of devices with such an old modem. I updated the know issue list of the wiki for falcon, that should be enough.

Patch-set: 1
Status: abandoned