libsensors: akmdfs: Add mininum delay
The daemon compares the sampling rates of the accelerometer and
compass sensors and picks the fastest one. The compass sensor
doesn't work properly with delays smaller than 200ms, so if the
accelerometer is working at faster rates, the compass stops working
properly. Add an hard limit of 200ms to prevent issues.

Change-Id: I7cc80da46694a3ef99ffc4c242e629da7c011065