libshims: Wrap property_get() to fix EXIF data
If the value of ro.build.tags is "test-keys", the camera vendor HAL
sets as phone model for the EXIF data "Phone" rather than the value
of ro.product.model. Currently we hack the HAL so that the check on
ro.build.tags is skipped, but that can be incovenient. Instead of
doing that, wrap property_get() so that we always return
"release-keys" as value for ro.build.tags.

Change-Id: I0d64411362be40b6eb644205bef5e817b8df27f0