libshims: Add libshims_camera for proper EXIF data
If the value of ro.build.tags is not "release-keys", the vendor
camera HAL sets "Phone" as phone model rather than the value of
ro.product.model. We currently hack the vendor HAL to have proper
EXIF data with both test-keys and release-keys, however this is
not convenient since it requires to manually edit the library.
Instead of hacking the library, wrap property_get() so that we
always return "release-keys" for ro.build.tags.

Change-Id: I0d64411362be40b6eb644205bef5e817b8df27f0