CMActions: Don't crash after screen rotations
If we don't remove SwitchBar listeners, we will have two different
listeners after screen rotation, with one that depends on a no
longer existing activity. This will cause a crash as soon as the
SwitchBar status is toggled since the extra listener will try to
use getActivity(), which will be null. Remove the listener when
the view is destroyed to prevent crashes.

Change-Id: I79663397c34b71a71845a9d6f8841da60cdf8857