libsensors: peregrine: Report correct gyroscope range and resolution
Android expects the values for the gyroscope to be expressed in rad/s.
According to the sensor specs, the max range is +/-2000 deg/s, so
convert the currently reported value from deg/s to rad/s. Also, the
sensor specs report a typical sensitivity of 0.07 deg/s for the used
range, so report that as resolution.

Change-Id: I20bff8e2e5856f2dc2f138ee1b1f1b48b25d47e2