libsensors: peregrine: Report correct gyro range and resolution
Android expects the values for the gyroscope to be expressed in rad/s.
According to the sensor specs, the max range is +/-2000 deg/s, so
convert the currently reported value from deg/s to rad/s. Also, the
sensor has a resolution of 16 bits, so calculate the right value
for the fixed range of +/-2000 deg/s.

Change-Id: I20bff8e2e5856f2dc2f138ee1b1f1b48b25d47e2