sensors: falcon-peregrine: Handle delays properly
The accelerometer, the compass and the orientation sensor should all
report their current delay to the akm kernel driver so that the akm
userspace daemon can poll data at the right rate. Currently we update
the delays only when Android sets a new delay, but we should update
them also when sensors get enabled or disabled. Also, since we use
API 1.3 for the sensors HAL delays can change even when the sensors
are disabled. This is probably not something the akm daemon expects,
so set report the current delay only if enabled and zero otherwise.
Lastly, limit the delay of the accelerometer to the maximum value
supported by the compass sensor when reporting the rate to the kernel.
This limit was enforced by the open source akm daemon, but since we
are back to the prebuilt daemon we must adjust the rate reported.

Change-Id: If505e37d004aae4d0654abf4c268aa35e0ac25da