sensors: falcon-peregrine: Initialize the accuracy of the sensors
The accelerometer and the gyroscope never change their accuracy
value. If we don't specify an initial value, they'll use an undefined
one and apps might constantly ask for a calibration even if not
required. Initialize also the accuracy of the compass and orientation
sensor, even if they should receive events for it. While at it, zero
out the entire sensor data structure rather than just part of it.

Change-Id: Id619667070e9287418d17a2ae9541ad016d22b47