Revert "sensors: falcon-peregrine: Query compass accuracy on init"
This was mostly needed due to the issue with the calibration now
solved in the open source akm daemon. The open source HAL from
AKM initializes the sensors with high accuracy, so why not do the
same here? If the sensor needs a calibration, we will receive an
event. The daemon can tell immediately if the sensor is not
calibrated, but it takes a while to confirm the opposite. By
starting with an high accuracy we will prevent the initial and
often unnecessary "figure-8" movement to confirm the initial
calibration.

This reverts commit d29cf805c4c578070c12f165bed969e4cca87cd1.

Change-Id: I8274733d2b67d8023aca6d38d53ccd62c669fa2b