sensors: akmdfs: Initialize sensor with low accuracy
The daemon is currently using only two accuracy values: 0 (unreliable)
and 3 (high). The accuracy is always set to 0 when the sensor is
enabled, regardless of the actual accuracy, and set to 3 after the
calibration that the user has to perform via the figure-8 movement.
Since the daemon can tell immediately if the sensor is uncalibrated
and since the sensor rarely needs to be calibrated to approximately
report the current orientation, use 1 (low) as default accuracy.
This allows apps that don't require the maximum accuracy to work
without requiring a calibration.

Change-Id: I87448c7e80bd1fea7581f148966705f14df372ff