sensors: akmdfs: Initialize sensor with low accuracy
The daemon is currently using two different accuracy values: 0
(unreliable) and 3 (high). Setting the initial value to 0 causes
some apps to constantly require a calibration via the figure-8
movement. This is usually unnecessary since the sensor is often
calibrated. The daemon can also tell immediately if the sensor is
unreliable, but not if it has an high accuracy. Therefore it makes
sense to use 1 (low) as default accuracy value.

Change-Id: I87448c7e80bd1fea7581f148966705f14df372ff